requirements.txt
python-telegram-bot
python-dotenv
spotipy
py-deezer
unidecode
eyed3
beautifulsoup4